+++ README.md
@@ -1,0 +1,5 @@
+# Revm - zkSync Era translation layer
+
+This crate acts as a bridge between Revm (Rust EVM implementation) and zkSync Era VM.
+
+Currenlty this is used by Foundry testing.
